Integre medios_contacto a personas
En un error en el merge volví a meter la clase fallecido.
La borré tal y como lo había hecho @gnoq.
Borre el modulo medios_contacto y metí todo en el modulo personas.
   y le puse un foreignkey a persona, con lo cual podríamos cargar
   los medios de contacto a los familiares y podríamos luego hacer
   un metodo para obtener las direcciones de cada bombero, de
   cada persona y de cada familiar de bombero.
