--- medios_contacto/admin.py
@@ -1,76 +1,0 @@
-from django.contrib import admin
-from django.utils.translation import ugettext_lazy as _
-from medios_contacto.models import (
-    DireccionPostal, DireccionWeb, Telefono, DireccionElectronica)
-
-
-@admin.register(DireccionPostal)
-class DireccionPostalAdmin(admin.ModelAdmin):
-    actions_on_bottom = True
-    fieldsets = (
-        (None, {
-            'fields': (
-                'uso',
-                'calle',
-                'numero',
-                'piso',
-                'departamento',
-                'localidad')
-            }),
-        (_("Observaciones"), {
-            'classes': ('collapse',),
-            'fields': ('observaciones',),
-            }))
-    search_fields = (
-        'calle',
-        'numero',
-        'piso',
-        'departamento',
-        'localidad__nombre')
-    list_filter = (
-        'calle',
-        'localidad')
-
-
-@admin.register(DireccionWeb)
-class DireccionWebAdmin(admin.ModelAdmin):
-    actions_on_bottom = True
-    fieldsets = (
-        (None, {
-            'fields': (
-                'tipo',
-                'direccion')
-            }),
-        (_("Observaciones"), {
-            'classes': ('collapse',),
-            'fields': ('observaciones',),
-            }))
-
-
-@admin.register(Telefono)
-class TelefonoAdmin(admin.ModelAdmin):
-    actions_on_bottom = True
-    fieldsets = (
-        (None, {
-            'fields': (
-                'tipo',
-                'telefono')
-            }),
-        (_("Observaciones"), {
-            'classes': ('collapse',),
-            'fields': ('observaciones',),
-            }))
-
-
-@admin.register(DireccionElectronica)
-class DireccionElectronicaAdmin(admin.ModelAdmin):
-    actions_on_bottom = True
-    fieldsets = (
-        (None, {
-            'fields': (
-                'mail',)
-            }),
-        (_("Observaciones"), {
-            'classes': ('collapse',),
-            'fields': ('observaciones',),
-            }))
