--- medios_contacto/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-from django.utils.translation import ugettext_lazy as _
-
-
-class MediosContactoConfig(AppConfig):
-    name = 'medios_contacto'
-    verbose_name = _("Medios de Contacto")
