--- medios_contacto/choices.py
@@ -1,17 +1,0 @@
-from django.utils.translation import ugettext_lazy as _
-
-
-USO_MEDIO = (
-    ('', _("Particular")),
-    ('L', _("Laboral")),
-)
-
-TIPO_WEB = (
-    ('', _("Perfil Web Social")),
-    ('L', _("Pagina Web Laboral")),
-)
-
-TIPO_TELEFONO = (
-    ('', _("Celular")),
-    ('F', _("Fijo"))
-)
