+++ personas/admin.py
@@ -1,6 +1,12 @@
 from django.contrib import admin
-from personas.models import Persona, Bombero, Fallecido
 from django.utils.translation import ugettext as _
+from personas.models import (
+    Persona,
+    Bombero,
+    DireccionPostal,
+    DireccionWeb,
+    Telefono,
+    DireccionElectronica)
 
 
 @admin.register(Persona)
@@ -76,23 +82,77 @@ class BomberoAdmin(admin.ModelAdmin):
     date_hierarchy = 'fecha_nacimiento'
 
 
-@admin.register(Fallecido)
-class FallecidoAdmin(admin.ModelAdmin):
+@admin.register(DireccionPostal)
+class DireccionPostalAdmin(admin.ModelAdmin):
     actions_on_bottom = True
-    list_display = (
-        'apellido',
-        'nombre',
-        'tipo_documento',
-        'documento',
-        'fecha_nacimiento',
-        'fecha_desceso')
+    fieldsets = (
+        (None, {
+            'fields': (
+                'persona',
+                'uso',
+                'calle',
+                'numero',
+                'piso',
+                'departamento',
+                'localidad')
+            }),
+        (_("Observaciones"), {
+            'classes': ('collapse',),
+            'fields': ('observaciones',),
+            }))
     search_fields = (
-        'apellido',
-        'nombre',
-        'documento',)
+        'calle',
+        'numero',
+        'piso',
+        'departamento',
+        'localidad__nombre')
     list_filter = (
-        'apellido',
-        'fecha_nacimiento',
-        'tipo_documento',
-        'fecha_desceso')
-    date_hierarchy = 'fecha_desceso'
+        'calle',
+        'localidad')
+
+
+@admin.register(DireccionWeb)
+class DireccionWebAdmin(admin.ModelAdmin):
+    actions_on_bottom = True
+    fieldsets = (
+        (None, {
+            'fields': (
+                'persona',
+                'tipo',
+                'direccion')
+            }),
+        (_("Observaciones"), {
+            'classes': ('collapse',),
+            'fields': ('observaciones',),
+            }))
+
+
+@admin.register(Telefono)
+class TelefonoAdmin(admin.ModelAdmin):
+    actions_on_bottom = True
+    fieldsets = (
+        (None, {
+            'fields': (
+                'persona',
+                'tipo',
+                'telefono')
+            }),
+        (_("Observaciones"), {
+            'classes': ('collapse',),
+            'fields': ('observaciones',),
+            }))
+
+
+@admin.register(DireccionElectronica)
+class DireccionElectronicaAdmin(admin.ModelAdmin):
+    actions_on_bottom = True
+    fieldsets = (
+        (None, {
+            'fields': (
+                'persona',
+                'mail',)
+            }),
+        (_("Observaciones"), {
+            'classes': ('collapse',),
+            'fields': ('observaciones',),
+            }))
